Dunno

+++ src/models/listing.py
@@ -1,5 +1,5 @@
 import uuid
-from sqlalchemy import UUID, String, DECIMAL, Integer, Text, DateTime, func, Boolean, ForeignKey, ARRAY, Enum
+from sqlalchemy import UUID, Float, String, DECIMAL, Integer, Text, DateTime, func, Boolean, ForeignKey, ARRAY, Enum
 from sqlalchemy.orm import mapped_column, relationship
 
 from .base import Base
@@ -27,15 +27,15 @@ class PropertyListing(Base):
     amenities = mapped_column(ARRAY(Enum(Amenities.INTERNET, Amenities.AC, Amenities.SECURITY, Amenities.POWER, name="amenities_enum")), default=[])
     location = mapped_column(String(100), index=True)
     google_maps_address = mapped_column(String(500))
-    latitude = mapped_column(DECIMAL(9, 6), nullable=True)
-    longitude = mapped_column(DECIMAL(9, 6), nullable=True)
-    price_per_night = mapped_column(DECIMAL(10, 2), index=True)
+    latitude = mapped_column(Float, nullable=True)
+    longitude = mapped_column(Float, nullable=True)
+    price_per_night = mapped_column(Float, index=True)
     room_count = mapped_column(Integer, default=0, index=True)
     listing_area = mapped_column(Integer, default=0)
     bathrooms_count = mapped_column(Integer, default=0)
     description = mapped_column(Text)
     total_reviews = mapped_column(Integer, default=0)
-    average_rating = mapped_column(DECIMAL, default=0)
+    average_rating = mapped_column(Float, default=0)
     is_available = mapped_column(Boolean, default=True, index=True)
     created_at = mapped_column(DateTime, server_default=func.now())
     updated_at = mapped_column(DateTime, server_default=func.now(), onupdate=func.now())
